requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly.express
numpy
scikit-learn
streamlit-option-menu